pyproject.toml
[project]
name = "embs"
version = "0.1.0"
description = "日本語テキストのセマンティック検索CLIツール"
requires-python = ">=3.11"
dependencies = [
    "typer",
    "docling",
    "sentence-transformers",
    "sqlite-vec",
    "atlassian-python-api",
]

[project.scripts]
embs = "embs.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
